% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habit_success_rate.R
\name{habit_success_rate}
\alias{habit_success_rate}
\title{Calculate the success rate for a habit}
\usage{
habit_success_rate(habit)
}
\arguments{
\item{habit}{A habit object.}
}
\value{
Success rate as a percentage.
}
\description{
Calculate the success rate for a habit
}
\examples{
# Create a habit, log a few days, and check success rate
habit2 <- add_habit("Exercise", "daily")
habit2 <- log_habit(habit2, lubridate::today() - 2)
habit2 <- log_habit(habit2, lubridate::today())
habit_success_rate(habit2)
}
