% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_habit.R
\name{log_habit}
\alias{log_habit}
\title{Log a completed habit}
\usage{
log_habit(habit, date = lubridate::today())
}
\arguments{
\item{habit}{A habit object.}

\item{date}{The date the habit was completed (defaults to current day).}
}
\value{
Updated habit object.
}
\description{
Log a completed habit
}
\examples{
Create and log a habit (example):
habit1 <- add_habit("Read Book", "daily")
habit1 <- log_habit(habit1)
print(habit1)
}
