% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habit_class.R
\name{new_habit}
\alias{new_habit}
\title{Create a new habit}
\usage{
new_habit(name, frequency)
}
\arguments{
\item{name}{Name of the habit.}

\item{frequency}{Target frequency (e.g., "daily", "weekly").}
}
\value{
A habit object.
}
\description{
Create a new habit
}
\examples{
How to create a new habit manually:
example_habit <- new_habit("Exercise", "daily")
print(example_habit)
}
